# Add specialized tablet_sstable_set