--- a/replica/database.hh
+++ b/replica/database.hh
@@ -1240,6 +1240,8 @@ public:
     friend class compaction_group;
 };
 
+lw_shared_ptr<sstables::sstable_set> make_tablet_sstable_set(schema_ptr, const storage_group_manager& sgm, const locator::tablet_map&);
+
 using user_types_metadata = data_dictionary::user_types_metadata;
 
 using keyspace_metadata = data_dictionary::keyspace_metadata;