--- a/replica/table.cc
+++ b/replica/table.cc
@@ -118,11 +118,11 @@ table::make_sstable_reader(schema_ptr s,
     }
 }
 
-lw_shared_ptr<sstables::sstable_set> compaction_group::make_compound_sstable_set() {
+lw_shared_ptr<sstables::sstable_set> compaction_group::make_compound_sstable_set() const {
     return make_lw_shared(sstables::make_compound_sstable_set(_t.schema(), { _main_sstables, _maintenance_sstables }));
 }
 
-lw_shared_ptr<sstables::sstable_set> table::make_compound_sstable_set() {
+lw_shared_ptr<sstables::sstable_set> table::make_compound_sstable_set() const {
     if (auto cg = single_compaction_group_if_available()) {
         return cg->make_compound_sstable_set();
     }