… replica/tablets.cc
… scylla-gdb.py