@avikivity 2024-02-04
One thing we can do later, is have restricted versions of sstable_set that capture only a token range. Most sstable_sets will be happy to just capture a single tablet (repair, migration, BYPASS CACHE reads sent from storage_proxy). Only row_cache needs a full set.