@scylladb-promoter 2024-02-05
#### :red_circle: CI State: FAILURE

:x: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/6402/)
#### Build Details:
- Duration: 12 min
- Builder: spider1.cloudius-systems.com