@bhalevy 2024-02-05
In v2 ([15ad7e4](https://github.com/scylladb/scylladb/commit/15ad7e4c248c571b9a582626af28204899ff884d)):
- rebased
- tablet_storage_group_manager: precalculate my_host_id and _tablet_map
  -  pass e_r_m by ref (not using ref to e_r_m_ptr)
- table: move compaction_group_list and storage_group_vector to storage_group_manager
  - keep table& in derived storage_group_manager classes
  - move also `handle_tablet_split_completion` into `tablet_storage_group_manager`
- tablet_storage_group_manager: handle_tablet_split_completion: pre-calc new_tablet_count
  - cosmetic change only