@scylladb-promoter 2024-02-05
#### :red_circle: CI State: FAILURE

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/6411/)
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/6411/)
:white_check_mark: - [dtest](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release/7660/)
:x: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/6411/)
#### Failed Tests (6253/22447):
- [test_count_and_group_by_row](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_count_and_group_by_row) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_count_and_group_by_row)
- [test_count_and_group_by_partition](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_count_and_group_by_partition) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_count_and_group_by_partition)
- [test_aggregation_inf_nan](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_aggregation_inf_nan) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_aggregation_inf_nan)
- [test_avg_decimal_2](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_avg_decimal_2) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_avg_decimal_2)
- [test_sum_in_partition](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_sum_in_partition) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_sum_in_partition)
- [test_sum_star](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_sum_star) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_sum_star)
- [test_count_and_group_by_row](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_count_and_group_by_row_2) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_count_and_group_by_row)
- [test_count_and_group_by_partition](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_count_and_group_by_partition_2) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_count_and_group_by_partition)
- [test_aggregation_inf_nan](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_aggregation_inf_nan_2) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_aggregation_inf_nan)
- [test_avg_decimal_2](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/6411/testReport/junit/%28root%29/test_aggregate/Tests___Unit_Tests___test_avg_decimal_2_2) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_avg_decimal_2)
#### Build Details:
- Duration: 2 hr 31 min
- Builder: spider3.cloudius-systems.com