@bhalevy 2024-02-06
In v3 ([1ef8af9](https://github.com/scylladb/scylladb/commit/1ef8af95179ee9eb76b9967d752995ef252219ad)):
- rebased + retested
- table: coroutinize update_effective_replication_map
  - wait of deregistration of empty main compaction groups after split instead of deferring it to a background fiber
- table: move compaction_group_list and storage_group_vector to storage_group_manager
  - dropped `mutable` from _compaction_groups member
  - moved additional compaction-group related helpers into storage_group_manager to reduce the exposure of compaction_groups() even further.
- table: tablet_storage_group_manager: storage_group_of: do not validate in release build mode