@raphaelsc 2024-02-06
The series look good to me. I am very tired today, so tomorrow morning I plan to review the incremental selection path.