@xemul 2024-02-09
For future references:
| branch | sct test id | test |
|------------|------------------------------------------------------|--------------------------------------------|
| master | 8d7cb3d2-1f22-408c-8b8e-9453a566d1cc | test_latency_write_with_nemesis |
| master | 87185741-7cb0-4f4a-9656-4d0d7cc084e0 | test_latency_read_with_nemesis |
| master | ce5d62ee-2ed6-4db3-af39-dec050286151 | test_latency_mixed_with_nemesis |
| this pr  | b6153834-6ceb-44a5-adc3-2be3e318f94d | test_latency_write_with_nemesis |
| this pr  | 099b1ad4-1c22-46af-be26-66b8df6e2b91 | test_latency_read_with_nemesis |
| this pr  | b265335c-9aa1-4507-8d40-083525d77702 | test_latency_mixed_with_nemesis |