@bhalevy 2024-02-11
in v5 ([4bba763](https://github.com/scylladb/scylladb/commit/4bba7636cb1d478f3dbbd52efea8e25a8b0bb99f)):
- rebased
- table: coroutinize update_effective_replication_map
  - collect and return unready futures from `compaction_group.stop()` in `handle_tablet_split_completion` so not to yield
  - awaited in update_effective_replication_map
- table: tablet_storage_group_manager: make tablet_sstable_set
  - instantiate a tablet_map in tablet_sstable_set to map token <-> tablet_id back and forth
  - tablet_incremental_select: invalidate _cur_selector if given `pos` if out of range for it
  - handle pos == dht::maximum case
  - rename `lowest_next_position` variable to `next_position` and improved comments around it