@bhalevy 2024-03-06
Latest benchmark results with cassandra-stress (will also summarize in this PR's description):

Before (0a7854ea4d)
```
$ sudo cpupower frequency-set -g userspace
$ scylla-developer-mode --smp=16 --memory=8G --experimental-features=consistent-topology-changes --experimental-features=tablets
cqlsh> CREATE KEYSPACE keyspace1 WITH replication={'class':'NetworkTopologyStrategy', 'replication_factor':1} AND tablets={'initial':2048};
$ ./tools/java/tools/bin/cassandra-stress write no-warmup n=10000000 -pop 'seq=1...10000000' -rate threads=128
Op rate                   :   76,806 op/s  [WRITE: 76,806 op/s]
Partition rate            :   76,806 pk/s  [WRITE: 76,806 pk/s]
Row rate                  :   76,806 row/s [WRITE: 76,806 row/s]
Latency mean              :    1.6 ms [WRITE: 1.6 ms]
Latency median            :    1.7 ms [WRITE: 1.7 ms]
Latency 95th percentile   :    3.0 ms [WRITE: 3.0 ms]
Latency 99th percentile   :    4.1 ms [WRITE: 4.1 ms]
Latency 99.9th percentile :    5.5 ms [WRITE: 5.5 ms]
Latency max               :   20.0 ms [WRITE: 20.0 ms]
Total partitions          : 10,000,000 [WRITE: 10,000,000]
Total errors              :          0 [WRITE: 0]
Total GC count            : 0
Total GC memory           : 0.000 KiB
Total GC time             :    0.0 seconds
Avg GC time               :    NaN ms
StdDev GC time            :    0.0 ms
Total operation time      : 00:02:10

$ scylla-api-client system drop_sstable_caches POST
$ ./tools/java/tools/bin/cassandra-stress read no-warmup duration=60s -pop 'dist=uniform(1..10000000)' -rate threads=128
Partition rate            :   34,330 pk/s  [READ: 34,330 pk/s]
Row rate                  :   34,330 row/s [READ: 34,330 row/s]
Latency mean              :    3.7 ms [READ: 3.7 ms]
Latency median            :    3.5 ms [READ: 3.5 ms]
Latency 95th percentile   :    7.0 ms [READ: 7.0 ms]
Latency 99th percentile   :    9.1 ms [READ: 9.1 ms]
Latency 99.9th percentile :   12.1 ms [READ: 12.1 ms]
Latency max               :   34.7 ms [READ: 34.7 ms]
Total partitions          :  2,070,374 [READ: 2,070,374]
Total errors              :          0 [READ: 0]
Total GC count            : 0
Total GC memory           : 0.000 KiB
Total GC time             :    0.0 seconds
Avg GC time               :    NaN ms
StdDev GC time            :    0.0 ms
Total operation time      : 00:01:00

$ scylla-api-client system drop_sstable_caches POST
$ ./tools/java/tools/bin/cassandra-stress mixed no-warmup duration=60s -pop 'dist=uniform(1..10000000)' -rate threads=128
Op rate                   :   32,195 op/s  [READ: 16,066 op/s, WRITE: 16,128 op/s]
Partition rate            :   32,195 pk/s  [READ: 16,066 pk/s, WRITE: 16,128 pk/s]
Row rate                  :   32,195 row/s [READ: 16,066 row/s, WRITE: 16,128 row/s]
Latency mean              :    3.9 ms [READ: 5.2 ms, WRITE: 2.7 ms]
Latency median            :    3.5 ms [READ: 4.8 ms, WRITE: 2.7 ms]
Latency 95th percentile   :    8.5 ms [READ: 10.1 ms, WRITE: 4.9 ms]
Latency 99th percentile   :   12.1 ms [READ: 13.7 ms, WRITE: 6.1 ms]
Latency 99.9th percentile :   17.4 ms [READ: 19.2 ms, WRITE: 7.9 ms]
Latency max               :   36.3 ms [READ: 36.3 ms, WRITE: 22.6 ms]
Total partitions          :  1,947,575 [READ: 971,911, WRITE: 975,664]
Total errors              :          0 [READ: 0, WRITE: 0]
Total GC count            : 0
Total GC memory           : 0.000 KiB
Total GC time             :    0.0 seconds
Avg GC time               :    NaN ms
StdDev GC time            :    0.0 ms
Total operation time      : 00:01:00
```

After (0c2c00f01b)
```
$ sudo cpupower frequency-set -g userspace
$ scylla-developer-mode --smp=16 --memory=8G --experimental-features=consistent-topology-changes --experimental-features=tablets
cqlsh> CREATE KEYSPACE keyspace1 WITH replication={'class':'NetworkTopologyStrategy', 'replication_factor':1} AND tablets={'initial':2048};
$ ./tools/java/tools/bin/cassandra-stress write no-warmup n=10000000 -pop 'seq=1...10000000' -rate threads=128
Op rate                   :  100,787 op/s  [WRITE: 100,787 op/s]
Partition rate            :  100,787 pk/s  [WRITE: 100,787 pk/s]
Row rate                  :  100,787 row/s [WRITE: 100,787 row/s]
Latency mean              :    1.3 ms [WRITE: 1.3 ms]
Latency median            :    1.1 ms [WRITE: 1.1 ms]
Latency 95th percentile   :    2.6 ms [WRITE: 2.6 ms]
Latency 99th percentile   :    3.1 ms [WRITE: 3.1 ms]
Latency 99.9th percentile :    4.4 ms [WRITE: 4.4 ms]
Latency max               :   25.6 ms [WRITE: 25.6 ms]
Total partitions          : 10,000,000 [WRITE: 10,000,000]
Total errors              :          0 [WRITE: 0]
Total GC count            : 0
Total GC memory           : 0.000 KiB
Total GC time             :    0.0 seconds
Avg GC time               :    NaN ms
StdDev GC time            :    0.0 ms
Total operation time      : 00:01:39

$ scylla-api-client system drop_sstable_caches POST
$ ./tools/java/tools/bin/cassandra-stress read no-warmup duration=60s -pop 'dist=uniform(1..10000000)' -rate threads=128
Op rate                   :  106,099 op/s  [READ: 106,099 op/s]
Partition rate            :  106,099 pk/s  [READ: 106,099 pk/s]
Row rate                  :  106,099 row/s [READ: 106,099 row/s]
Latency mean              :    1.2 ms [READ: 1.2 ms]
Latency median            :    1.1 ms [READ: 1.1 ms]
Latency 95th percentile   :    2.3 ms [READ: 2.3 ms]
Latency 99th percentile   :    3.1 ms [READ: 3.1 ms]
Latency 99.9th percentile :    4.7 ms [READ: 4.7 ms]
Latency max               :   68.2 ms [READ: 68.2 ms]
Total partitions          :  6,374,957 [READ: 6,374,957]
Total errors              :          0 [READ: 0]
Total GC count            : 0
Total GC memory           : 0.000 KiB
Total GC time             :    0.0 seconds
Avg GC time               :    NaN ms
StdDev GC time            :    0.0 ms
Total operation time      : 00:01:00

$ scylla-api-client system drop_sstable_caches POST
$ ./tools/java/tools/bin/cassandra-stress mixed no-warmup duration=60s -pop 'dist=uniform(1..10000000)' -rate threads=128
Op rate                   :   79,246 op/s  [READ: 39,648 op/s, WRITE: 39,598 op/s]
Partition rate            :   79,246 pk/s  [READ: 39,648 pk/s, WRITE: 39,598 pk/s]
Row rate                  :   79,246 row/s [READ: 39,648 row/s, WRITE: 39,598 row/s]
Latency mean              :    1.6 ms [READ: 1.8 ms, WRITE: 1.3 ms]
Latency median            :    1.6 ms [READ: 1.9 ms, WRITE: 1.2 ms]
Latency 95th percentile   :    3.3 ms [READ: 3.5 ms, WRITE: 2.5 ms]
Latency 99th percentile   :    4.0 ms [READ: 4.2 ms, WRITE: 3.0 ms]
Latency 99.9th percentile :    5.5 ms [READ: 6.0 ms, WRITE: 4.3 ms]
Latency max               :   42.5 ms [READ: 42.5 ms, WRITE: 39.9 ms]
Total partitions          :  4,819,553 [READ: 2,411,286, WRITE: 2,408,267]
Total errors              :          0 [READ: 0, WRITE: 0]
Total GC count            : 0
Total GC memory           : 0.000 KiB
Total GC time             :    0.0 seconds
Avg GC time               :    NaN ms
StdDev GC time            :    0.0 ms
Total operation time      : 00:01:00
```